general cleaning and ordering
