+++ .gitignore
@@ -73,6 +73,8 @@ db.sqlite3-journal
 
 # Flask stuff:
 instance/
+migrations/
+!app/migrations/
 
 # Scrapy stuff:
 .scrapy
